feat: implement particle space animation and integrate into background page
