+++ hooks/useParticleSpaceAnimation.ts
@@ -1,0 +1,144 @@
+import { useEffect, useRef } from "react";
+
+export const useParticleSpaceAnimation = () => {
+  const blurCanvasRef = useRef<HTMLCanvasElement | null>(null);
+  const charCanvasRef = useRef<HTMLCanvasElement | null>(null);
+
+  useEffect(() => {
+    if (typeof window === 'undefined') return;
+    const blurCanvas = blurCanvasRef.current;
+    const charCanvas = charCanvasRef.current;
+    if (!blurCanvas || !charCanvas) return;
+
+    const blurCtx = blurCanvas.getContext('2d');
+    const charCtx = charCanvas.getContext('2d');
+    if (!blurCtx || !charCtx) return;
+
+    const BGCOLOR = '#121212';
+    const FADE_ALPHA = 0.08;
+    const COUNT = 12;
+    const FOCUS_DISTANCE = 150;
+    const DEPTH = 500;
+    const SPEED = 1.5;
+    const GLITCH_PROB = 0.98;
+    const CHAR_SIZE = 20;
+
+    document.documentElement.style.backgroundColor = BGCOLOR;
+    document.body.style.backgroundColor = BGCOLOR;
+    blurCanvas.style.backgroundColor = BGCOLOR;
+    charCanvas.style.backgroundColor = 'transparent';
+
+    const viewport = { width: window.innerWidth, height: window.innerHeight };
+    const rand = (range: number) => (Math.random() - 0.5) * range;
+    const withAlpha = (hex: string, alpha: number) => {
+      const value = Math.round(Math.min(Math.max(alpha, 0), 1) * 255)
+        .toString(16)
+        .padStart(2, '0');
+      return `${hex}${value}`;
+    };
+
+    class Particle {
+      x = 0;
+      y = 0;
+      z = 0;
+      ch = '0';
+      doBug = 0;
+
+      constructor() {
+        this.reset();
+      }
+
+      reset() {
+        this.x = rand(viewport.width);
+        this.y = rand(viewport.height);
+        this.z = Math.random() * DEPTH;
+        this.ch = Math.random() > 0.5 ? '0' : '1';
+        this.doBug = 0;
+      }
+
+      update() {
+        if ((this.z -= SPEED) < 1) this.reset();
+
+        if (Math.random() < 0.0001) this.doBug = Math.floor(Math.random() * 330) + 30;
+        if (this.doBug) {
+          this.reset();
+          this.doBug--;
+        }
+      }
+
+      draw() {
+        const scale = FOCUS_DISTANCE / (FOCUS_DISTANCE + this.z);
+        const x = (this.x * scale) + (viewport.width * 0.5);
+        const y = (this.y * scale) + (viewport.height * 0.5);
+
+        if (x < 0 || x > viewport.width || y < 0 || y > viewport.height) return;
+
+        const size = Math.max(1, CHAR_SIZE * scale);
+        const alpha = Math.min(1, Math.max(0.1, (DEPTH - this.z) / DEPTH));
+        const chars = [{ color: withAlpha('#1eff28', alpha), x, y }];
+
+        if (Math.random() > GLITCH_PROB) {
+          const offsetX = rand(16);
+          const offsetY = rand(12);
+          chars.push({ color: withAlpha('#ff3c3c', alpha), x: x + offsetX, y: y + offsetY });
+          chars.push({ color: withAlpha('#5ae6ff', alpha), x: x - offsetX, y: y - offsetY });
+        }
+
+        const font = `${size}px Fira Code, monospace`;
+        blurCtx.font = font;
+        charCtx.font = font;
+        chars.forEach(({ color, x: cx, y: cy }) => {
+          blurCtx.fillStyle = color;
+          blurCtx.fillText(this.ch, cx, cy);
+          charCtx.fillStyle = color;
+          charCtx.fillText(this.ch, cx, cy);
+        });
+      }
+    }
+
+    const resize = () => {
+      viewport.width = window.innerWidth;
+      viewport.height = window.innerHeight;
+      blurCanvas.width = viewport.width;
+      blurCanvas.height = viewport.height;
+      charCanvas.width = viewport.width;
+      charCanvas.height = viewport.height;
+      blurCtx.fillStyle = BGCOLOR;
+      blurCtx.fillRect(0, 0, viewport.width, viewport.height);
+      charCtx.clearRect(0, 0, viewport.width, viewport.height);
+      particles.forEach(p => p.reset());
+    };
+
+    const fadeBlur = () => {
+      blurCtx.fillStyle = `rgba(18, 18, 18, ${FADE_ALPHA})`;
+      blurCtx.fillRect(0, 0, viewport.width, viewport.height);
+    };
+
+    const clearCharLayer = () => {
+      charCtx.clearRect(0, 0, viewport.width, viewport.height);
+    };
+
+    let animationFrameId = 0;
+    const animate = () => {
+      fadeBlur();
+      clearCharLayer();
+      particles.forEach(p => {
+        p.update();
+        p.draw();
+      });
+      animationFrameId = window.requestAnimationFrame(animate);
+    };
+
+    const particles = Array.from({ length: COUNT }, () => new Particle());
+
+    window.addEventListener('resize', resize);
+    resize();
+    animate();
+
+    return () => {
+      window.removeEventListener('resize', resize);
+      window.cancelAnimationFrame(animationFrameId);
+    };
+  }, []);
+    return {blurCanvasRef, charCanvasRef};
+};
