+++ pages/api/NextPageWithLayout.ts
@@ -1,0 +1,6 @@
+import type { NextPage } from "next";
+import type { ReactElement, ReactNode } from "react";
+
+export type NextPageWithLayout<P = {}> = NextPage<P> & {
+  getLayout?: (page: ReactElement) => ReactNode;
+};
