+++ public/welcome.md
@@ -1,9 +1,7 @@
 # 電子計算研究部
 ようこそ、大阪産業大学 文化会 電子計算研究部のウェブサイトへ！愛称は「電研」です。
-<br>
 パソコン、ネットワーク、ゲーム、プログラミングといった活動を通じて、技術力の向上や部員同士との交流を深めています。
 <br>
-<!-- また、初心者から経験者まで幅広く募集しています。興味のある方は途中参加もできますので、ぜひ参加してください。新入生の方は4月の新入生歓迎会もお見逃しなく！そして毎年、大学祭にて展示を行っていますので気軽にお越しください！<br> -->
 
 ## お知らせ
 - 2025.04.01: 新入生歓迎会を開催しました
