+++ styles/Home.module.css
@@ -108,7 +108,8 @@
   text-align: left;
   color: inherit;
   text-decoration: none;
-  background-color: #ffffff10;
+  background-color: #2f2f2f8f;
+  backdrop-filter: blur(1px);
   border: 1px solid #ffffff1f;
   border-radius: 10px;
   transition: color 0.15s ease, border-color 0.15s ease;
@@ -148,13 +149,13 @@
 
 .terminal {
   width: 640px;
-  height: 340px;
+  height: 320px;
   margin-top: 20px;
   margin-bottom: 12px;
-
   overflow-x: hidden;
   overflow-y: scroll;
-  background-color: #ffffff10;
+  background-color: #2f2f2f8f;
+  backdrop-filter: blur(1px);
   border: 1px solid #ffffff1f;
   border-radius: 10px;
 }
